feat: reorganize DTO and keys structure; migrate to shared/data directory
