+++ src/shared/data/DTO/AttributesDTO.ts
@@ -1,4 +1,4 @@
-import { AttributeKey, ATTR_KEYS } from "shared/keys";
+import { AttributeKey, ATTR_KEYS } from "shared/data/keys";
 
 export interface AttributesDTO {
 	[key: string]: number;
