+++ src/shared/data/DTO/FillBarBaseDTO.ts
