+++ src/shared/data/DTO/ResourceDTO.ts
