+++ src/shared/data/DTO/index.ts
@@ -1,2 +1,3 @@
 export * from "./ResourceDTO";
 export * from "./FillBarBaseDTO";
+export * from "./ResourceDTO";
