+++ src/shared/data/keys/index.ts
@@ -1,0 +1,1 @@
+export * from "./attributes";
